refactor: move KV-transfer enums to backend-agnostic header
Move TransferDirection and EngineKVFormat enums from
mem_kernels.cuh (CUDA) and mem_kernels_sycl.h (SYCL)
into a new shared header csrc/kv_transfer_types.h.

This avoids duplicating these enums in every new backend
and is a first step toward full refactor.

Part of #3676
